fixes based on comments to GH1020, including adding separate excepptions module for better import to multiple modules
